backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.3
sendgrid==6.11.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.3
requests==2.31.0
passlib[bcrypt]
PyJWT
